quell warning
git-svn-id: https://svn.code.sf.net/p/bzflag/code/trunk/bzflag@4344 08b3d480-bf2c-0410-a26f-811ee3361c24

+++ src/platform/MacOSX/MacWindow.cxx
@@ -390,7 +390,7 @@ MacWindow::MacWindow(const MacDisplay *display, MacVisual *visual) :
   /* set the size of the window to the default (existing resolution). this
    * includes initializing the opengl context to the correct aspect ratio.
    */
-  setSize(settings.Window_Size.width, settings.Window_Size.height);
+  setSize((int)settings.Window_Size.width, (int)settings.Window_Size.height);
 
   display->setContext(gl_context.GetGLContext());
